(PC-43131)[API] chore: delete or move /routes/shared
