+++ api/src/pcapi/core/educational/validation.py
@@ -9,6 +9,7 @@ from pcapi.core.educational import repository
 from pcapi.core.offers import models as offers_models
 from pcapi.models import api_errors
 from pcapi.models import db
+from pcapi.models.feature import FeatureToggle
 
 
 def check_institution_fund(booking_amount: Decimal, deposit: models.EducationalDeposit) -> None:
@@ -166,3 +167,48 @@ def check_contact_request(offer: models.CollectiveOfferTemplate, updates: dict) 
 
     if set_url and set_form:
         raise exceptions.UrlandFormBothSetError()
+
+
+def validate_students(student_levels: list[str] | list[models.StudentLevels] | None) -> list[models.StudentLevels]:
+    if not student_levels:
+        raise ValueError("La liste des niveaux scolaires ne peut pas être vide")
+
+    output = []
+    for student_level in student_levels:
+        try:
+            if isinstance(student_level, str):
+                level = _fetch_student_level(student_level)
+            else:
+                level = student_level
+
+        except ValueError:
+            permitted = _permitted_levels()
+            msg = f'Value is not a valid enumeration member; permitted: ["{permitted}"]'
+            raise ValueError(msg)
+
+        if not FeatureToggle.ENABLE_MARSEILLE.is_active() and level in models.StudentLevels.primary_levels():
+            raise ValueError(f"Ce niveau ('{level.name}') n'est pas encore autorisé")
+
+        output.append(level)
+    return output
+
+
+def _permitted_levels() -> str:
+    if FeatureToggle.ENABLE_MARSEILLE.is_active():
+        return '", "'.join(models.StudentLevels.__members__.keys())
+
+    restricted_levels = [
+        level.name for level in models.StudentLevels if level not in models.StudentLevels.primary_levels()
+    ]
+    return '", "'.join(restricted_levels)
+
+
+def _fetch_student_level(level: str) -> models.StudentLevels:
+    """Fetch student level from enum name or value (try both)"""
+    try:
+        return models.StudentLevels(level)
+    except ValueError:
+        try:
+            return models.StudentLevels[level]
+        except Exception:
+            raise ValueError("Unknown student level")
