+++ api/src/pcapi/routes/pro/__init__.py
@@ -7,6 +7,7 @@ def install_routes(app: Flask) -> None:
     from . import collective_bookings
     from . import collective_offers
     from . import collective_stocks
+    from . import educational_domains
     from . import educational_institutions
     from . import features
     from . import finance
