+++ api/src/pcapi/routes/pro/educational_domains.py
