+++ api/src/pcapi/routes/serialization/cookies_consent.py
