+++ api/tests/routes/pro/list_educational_domains_test.py
